Fix corrupted and out-of-order messages
The corrupted messages would happen occasionally and once it did, the player would stop receiving messages from the opponent. So while the opponent may have been moving, the player wouldn't see that. This should fix that, and maybe also the message ordering (seems so), so we shouldn't need the sequence numbers and priority queue to enforce the message ordering anymore.